Chat App E2EE Simple/requirements.txt
chardet==5.1.0
pyaes==1.6.1
pycryptodome==3.17
